+++ src/main/java/org/folio/rest/impl/InitAPIs.java
@@ -1,15 +1,20 @@
 package org.folio.rest.impl;
 
-import io.vertx.core.AsyncResult;
-import io.vertx.core.Context;
-import io.vertx.core.Future;
-import io.vertx.core.Handler;
-import io.vertx.core.Vertx;
+import io.vertx.core.*;
+import io.vertx.serviceproxy.ServiceBinder;
+import org.folio.rest.resource.interfaces.InitAPI;
+import org.folio.services.ConfigurationService;
+import org.folio.services.StorageService;
+
 import java.net.URL;
 import java.util.MissingResourceException;
-import org.folio.rest.resource.interfaces.InitAPI;
+
+import static org.folio.util.LoginConfigUtils.EVENT_CONFIG_PROXY_CONFIG_ADDRESS;
+import static org.folio.util.LoginConfigUtils.EVENT_CONFIG_PROXY_STORY_ADDRESS;
 
 /**
+ * Performs preprocessing operations before the verticle is deployed,
+ * e.g. components registration, initializing, binding.
  *
  * @author kurt
  */
@@ -21,8 +26,14 @@ public class InitAPIs implements InitAPI {
     if(u == null) {
       resultHandler.handle(Future.failedFuture(new MissingResourceException(CREDENTIAL_SCHEMA_PATH, InitAPIs.class.getName(), CREDENTIAL_SCHEMA_PATH)));
     } else {
+      new ServiceBinder(vertx)
+        .setAddress(EVENT_CONFIG_PROXY_STORY_ADDRESS)
+        .register(StorageService.class, StorageService.create(vertx));
+      new ServiceBinder(vertx)
+        .setAddress(EVENT_CONFIG_PROXY_CONFIG_ADDRESS)
+        .register(ConfigurationService.class, ConfigurationService.create(vertx));
+
       resultHandler.handle(Future.succeededFuture(true));
     }
   }
-
 }